medicinal-plant-detection/backend/requirements.txt
flask==2.3.2
flask-cors==3.0.10
numpy==1.24.3
pillow==9.5.0
torch==2.1.0
safetensors==0.4.3
werkzeug==2.3.7
